model-deployment/requirements.txt
streamlit
requests
opencv-python
numpy
Pillow